added i2c added Port map
i2c is only partly working at this time. Reading from the compass
(getHeading) causes a segfault. It has something to do with the number
of bytes being sent back to the rio and the size of the buffer.

The port map is complete, but some items under the expansion header
will change.